Refactor: remove redundant uint64() type convertion

+++ pkg/kubelet/cm/util/cgroups.go
@@ -19,11 +19,11 @@ package util
 // Convert cgroup v1 cpu.shares value to cgroup v2 cpu.weight
 // https://github.com/kubernetes/enhancements/tree/master/keps/sig-node/2254-cgroup-v2#phase-1-convert-from-cgroups-v1-settings-to-v2
 func CPUSharesToCPUWeight(cpuShares uint64) uint64 {
-	return uint64((((cpuShares - 2) * 9999) / 262142) + 1)
+	return (((cpuShares - 2) * 9999) / 262142) + 1
 }
 
 // Convert cgroup v2 cpu.weight value to cgroup v1 cpu.shares
 // https://github.com/kubernetes/enhancements/tree/master/keps/sig-node/2254-cgroup-v2#phase-1-convert-from-cgroups-v1-settings-to-v2
 func CPUWeightToCPUShares(cpuWeight uint64) uint64 {
-	return uint64((((cpuWeight - 1) * 262142) / 9999) + 2)
+	return (((cpuWeight - 1) * 262142) / 9999) + 2
 }
